Setup Godot engine in workflow
Download the stable version of Godot, using the info found in the
previous job. It is installed in ~/.local/share/applications so that it
becomes available to the runner from anywhere within it (i.e. in $PATH).

A soft link is created to simplify the name for later use, as versions
and architectures change.