Add caching of Godot
This will cache the Godot setup between runs so that it doesn't need to
be run on every workflow run. This one uses the Godot version and system
architecture as the cache key.

This cache will also mean Godot will be available to other jobs in the
workflow. This allows the workflow to be split into jobs, making the
workflow definition more concise.

The folder where the downloaded Godot zip file is changed to a different
folder. This is to make sure only the engine is cached, and not the
downloaded .zip file.