Save compiled Rust GDExtension code as artifact
Prior to this change, there was no way for the final build game job to
refer to the compiled Rust GDExtension game code. The building of the
Rust GDExtension code takes some time. It is good to have that job run
separately to other jobs to reduced the overall runtime of the workflow.

This change creates a GitHub workflow artifact of the compiled code.
This will allow the binary to be used when the game is built.

An artifact is used instead of a cache because we (assume) want to
rebuild the game code on every run. Caches are more suitable for no
changes between runs. Making a new cache each run would fill the cache
space. So artifact is a more use here.